Collect a memory track setup function in test fixtures
